+++ Sources/BraintreeCard/BTCardClient.swift
@@ -10,17 +10,17 @@ import BraintreeCore
     // MARK: - Internal Properties
 
     /// Exposed for testing to get the instance of BTAPIClient
-    let apiClient: BTAPIClient
+    var apiClient: BTAPIClient
 
     let graphQLTokenizeFeature: String = "tokenize_credit_cards"
 
     // MARK: - Initializer
 
     /// Creates a card client
-    /// - Parameter apiClient: An API client
-    @objc(initWithAPIClient:)
-    public init(apiClient: BTAPIClient) {
-        self.apiClient = apiClient
+    /// - Parameter authorization: A valid client token or tokenization key used to authorize API calls.
+    @objc(initWithAuthorization:)
+    public init(authorization: String) {
+        self.apiClient = BTAPIClient(newAuthorization: authorization)
     }
 
     // MARK: - Public Methods